requirements.txt
pyqt5
PyGLM
numpy
pillow
pyopengl